requirements.txt
#opencv_python==4.4.0.46
albumentations==0.5.1 --no-binary=imgaug,albumentations
ray>=1.0.1
einops==0.3.0
kornia==0.7.0
loguru==0.5.3
yacs>=0.1.8
tqdm
autopep8
pylint
ipython
jupyterlab
matplotlib
h5py
pytorch-lightning==1.3.5
loguru
joblib>=1.0.1
torchmetrics==0.4
pyproj